mobsf_custom_api/requirements_test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
requests-mock==1.11.0